20240111 @Mookse - add `alert.json` - rename modules => `.mjs`
