+++ inc/js/mylife-datasource-config.mjs
